requirements.txt
pycryptodome==3.19.0
cryptography==41.0.7
python-dotenv==1.0.0
